Tons of new things discovered (#27)
* Changes in enums and p4g.tbl

There are tons of enums added in affinity status. Example:

	Neutral5percentdamagereduced = 1,
	Neutral10percentdamagereduced = 2,
	Neutral15percentdamagereduced = 3,
	Neutral20percentdamagereduced = 4,
	Neutral25percentdamagereduced = 5, and etc,

Also, fixed wrong buffdebuff2 text and its' true effect => Physical and Magic shield are not what they are. Those are actually Hit and Evasion up/down and the latter one is actually dekaja and dekunda status.

* Tons of new things discovered

In p5_tbl, tons of bits are discovered under UNIT.TBL and ENCOUNT.TBL. There are comments to explain what the functions of the bits are for
In p5_struct, added a new function for Item category and item type in UNIT.TBL
In p5_enums, added new value functions that are not discovered before