Update p5r_tbl.bt
Learnability of a skill are now defined via an enum in p5r_tbl, with no skill being Nothing, a skill as Skill, and a trait as Trait. (mindblowing) The Party Member enum of it is exactly the same, just with a 0 at the end of them to avoid any errors. 
Also removes the comment beside it saying False=0, True=1, Trait=3 because its now said when looking at the value.