Update p3p_tbl.bt
Base Power section updated to show the correct number courtesy of Nerdu
The Member Status structure is now gone, as it interfered with updating the Ailments structure. 
Dizzy, Orgia Mode, and Overheat have been added to the Ailments structure. The other effects in there have been properly reordered.
The Ailments structure is now renamed to Effects, as it has more than just Ailments.