Reverting the ENCOUNT.TBL part of the PR as further research is needed
BlazingDiancie pointed out that the Reaper cannot be Critted even though it has none of the "No Critical" or "Prevent Knockdwon" flags set in its encounter or as a unit, so some other flag (or special code?) must be in play.  It doesn't make sense to update this table until we know for sure what affects crittability.