Convert unknown section of SKILL.TBL to Technical Map
The previously unknown section of SKILL.TBL has been renamed "Technical Map" because from my testing and from my analysis of the game code, this part of the table determines, for each ailment, which attacking affinities inflict Technical damage against combatants with that ailment.  I've included all of the structural information I was able to discern for this table section.