Add details about Affinities, Traits, and the Technical Combo Map (#65)
* Add Affinity details; fix Critical details

Adds info about affinities' effect on damage & ailment chances, the purpose of the first 2 bits of an affinity, and corrects a mistake in ENCOUNT.TBL regarding critical hits.

* Reverting the ENCOUNT.TBL part of the PR as further research is needed

BlazingDiancie pointed out that the Reaper cannot be Critted even though it has none of the "No Critical" or "Prevent Knockdwon" flags set in its encounter or as a unit, so some other flag (or special code?) must be in play.  It doesn't make sense to update this table until we know for sure what affects crittability.

* Touch up trait template

Add couple of facts about skills

* Convert unknown section of SKILL.TBL to Technical Map

The previously unknown section of SKILL.TBL has been renamed "Technical Map" because from my testing and from my analysis of the game code, this part of the table determines, for each ailment, which attacking affinities inflict Technical damage against combatants with that ailment.  I've included all of the structural information I was able to discern for this table section.

* Added technical damage multiplier field to Technical Combo Map